pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "shroudstone"
readme = "README.md"
authors = [{name = "Anthony Carapetis", email = "anthony.carapetis@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Topic :: Games/Entertainment :: Real Time Strategy"
]
keywords = ["Stormgate"]
dynamic = ["version", "description"]
dependencies = [
    "pillow>=5.0",
    "protobuf~=4.23",
    "pydantic~=2.6",
    "pystray~=0.19.5",
    "python-dateutil>=2.8.2",
    "pyyaml>=5",
    "typer[all]~=0.9.0",
    "typing_extensions>=4.7.1",
]
requires-python = "~=3.8"

[project.urls]
Source = "https://github.com/acarapetis/shroudstone"

[project.scripts]
shroudstone = "shroudstone.__main__:main"
